Move files to new dirs
