--- projects/task1/src/main.c
@@ -1,37 +1,0 @@
-#include <stdbool.h>
-#include <stdint.h>
-
-#include "FreeRTOS.h"
-#include "delay.h"
-#include "gpio.h"
-#include "log.h"
-#include "misc.h"
-#include "tasks.h"
-
-TASK(task1, TASK_STACK_512) {
-  int counter1 = 0;
-  while (true) {
-    LOG_DEBUG("Task 1 Started %d\n", counter1);
-    counter1++;
-    delay_ms(1000);
-  }
-}
-TASK(task2, TASK_STACK_512) {
-  int counter2 = 0;
-  while (true) {
-    LOG_DEBUG("Task 2 Started %d\n", counter2);
-    counter2++;
-    delay_ms(250);
-  }
-}
-int main(void) {
-  log_init();
-  tasks_init();
-
-  tasks_init_task(task1, TASK_PRIORITY(2), NULL);
-  tasks_init_task(task2, TASK_PRIORITY(3), NULL);
-  LOG_DEBUG("Program start...\n");
-
-  tasks_start();
-  return 0;
-}
